abstract out utils
